# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResBUSS": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"BussName": {
						"type": "string"
					},
					"BussSeats": {
						"type": "number",
						"format": "double"
					},
					"BussBokingDates": {
						"properties": {
							"endingDate": {
								"type": "string"
							},
							"startingDate": {
								"type": "string"
							}
						},
						"required": [
							"endingDate",
							"startingDate"
						],
						"type": "object"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"BussName",
					"BussSeats",
					"BussBokingDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetBUSS": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqBUSS": {
				"properties": {
					"BussName": {
						"type": "string"
					},
					"BussSeats": {
						"type": "number",
						"format": "double"
					},
					"BussBokingDates": {
						"properties": {
							"endingDate": {
								"type": "string"
							},
							"startingDate": {
								"type": "string"
							}
						},
						"required": [
							"endingDate",
							"startingDate"
						],
						"type": "object"
					}
				},
				"required": [
					"BussName",
					"BussSeats",
					"BussBokingDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqBUSS": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"BussName": {
						"type": "string"
					},
					"BussSeats": {
						"type": "number",
						"format": "double"
					},
					"BussBokingDates": {
						"properties": {
							"endingDate": {
								"type": "string"
							},
							"startingDate": {
								"type": "string"
							}
						},
						"required": [
							"endingDate",
							"startingDate"
						],
						"type": "object"
					}
				},
				"required": [
					"_id",
					"BussName",
					"BussSeats",
					"BussBokingDates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IBUSS": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteBUSS": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResTRIP": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Tripname": {
						"type": "string"
					},
					"participant_No": {
						"type": "number",
						"format": "double"
					},
					"busses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"CreatedAt": {
						"type": "string"
					},
					"UpdatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Tripname",
					"participant_No",
					"busses"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqTrip": {
				"properties": {
					"Tripname": {
						"type": "string"
					},
					"participant_No": {
						"type": "number",
						"format": "double"
					},
					"busses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"Tripname",
					"participant_No",
					"busses"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "final-task",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdul Wahab"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/buss/getbuss": {
			"post": {
				"operationId": "Getbuss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBUSS"
								}
							}
						}
					}
				},
				"tags": [
					"Buss"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetBUSS"
							}
						}
					}
				}
			}
		},
		"/buss/savebuss": {
			"post": {
				"operationId": "Savebuss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBUSS"
								}
							}
						}
					}
				},
				"tags": [
					"Buss"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqBUSS"
							}
						}
					}
				}
			}
		},
		"/buss/updatebuss": {
			"put": {
				"operationId": "Updatebuss",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBUSS"
								}
							}
						}
					}
				},
				"tags": [
					"Buss"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqBUSS"
							}
						}
					}
				}
			}
		},
		"/buss/deletebuss": {
			"delete": {
				"operationId": "Deletebuss",
				"responses": {
					"200": {
						"description": "Buss deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IBUSS"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Buss"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteBUSS"
							}
						}
					}
				}
			}
		},
		"/trip/savetrip": {
			"post": {
				"operationId": "SaveTrip",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResTRIP"
								}
							}
						}
					}
				},
				"summary": "Busses will automatically add through middle ware",
				"tags": [
					"Trip"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqTrip"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}